+++ src/main/java/net/turnbig/jdbcx/dialect/impl/OracleDialect.java
@@ -62,8 +62,8 @@ public class OracleDialect extends AbstractSQLDialect {
 	@Override
 	public String getPageableSql(String sql, Pageable pageable) {
 		String sortedSql = SelectSqlUtils.addSort(sql, pageable.getSort());
-		int startRow = pageable.getOffset();
-		int endRow = pageable.getOffset() + pageable.getPageSize();
+		long startRow = pageable.getOffset();
+		long endRow = pageable.getOffset() + pageable.getPageSize();
 		String pagedSql = "select * from ( select tmp_page.*, rownum row_id from ({0}) tmp_page "
 				+ " where rownum <= {1} ) where row_id > {2}";
 		return MessageFormat.format(pagedSql, sortedSql, endRow, startRow);
